+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Topcoder React Lib
 
-### v0.7.3
+### v0.7.4
 - Added `getTcM2mToken()` function to API service, to facilitate handling of
   TC M2M token;
-- Added `reduceGroupIds(..)` function to Groups service, to easily extract all
-  IDs from a group tree.
+- Added `getGroupTreeIds(..)`, `getTokenV3()`, and `reduceGroupIds(..)`
+  functions to the Groups service, to facilitate deduction of all IDs in a group
+  (sub-)tree, rooted at a specified group.
 
 ### v0.6.0
 A few changes related to fetch of projests visitble to users, and creation of